Split definitions into their own module
